Add prettier/eslint pre-commit step
**NOTE: If run into a git conflict due to this commit simply run `git checkout --ours -- <FILE>` to resolve the conflict with your changes and then prettier will re-run when you commit**

For people who are using prettier is important to make sure it runs before every commit and for every contributor. Otherwise there will be a lot of white-space noise in the commits of those who are running it.

This commit adds prettier and eslint to the pre-commit hook so all files will be formatted and linted before committing.